CAMP/CAMP/requirements.txt
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
cachelib==0.13.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
ConfigArgParse==1.7.1
cryptography==46.0.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
Flask-Session==0.8.0
gevent==25.9.1
geventhttpclient==2.3.5
greenlet==3.2.4
h11==0.16.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.42.5
locust-cloud==1.29.0
MarkupSafe==3.0.2
msgpack==1.1.2
msgspec==0.19.0
mysql-connector-python==9.4.0
mysqlclient==2.2.7
packaging==25.0
platformdirs==4.5.0
pluggy==1.6.0
psutil==7.1.3
pycparser==2.23
Pygments==2.19.2
pytest==9.0.1
python-engineio==4.12.3
python-socketio==5.14.3
pywin32==311
pyzmq==27.1.0
requests==2.32.4
simple-websocket==1.1.0
urllib3==2.5.0
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.3.2
zope.event==6.1
zope.interface==8.1.1
